+++ views/about.jade
@@ -29,7 +29,7 @@ block content
   p YASP only starts tracking your matches from when you signed in.  We get full match histories for contributors and active YASP users.
   
   h3 I signed in a while ago and just came back, why haven't my matches been tracked?
-  p To keep the load reasonable, we stop tracking players' matches after a period of inactivity.  Each visit to the YASP site while signed in resets this period.
+  p To keep the load reasonable, we stop parsing players' matches after a period of inactivity.  Each visit to the YASP site while signed in resets this period.
   
   h3 How do untracking and full history eligibility work?
   p Currently, we untrack users who have not visited YASP for 3 days.  For full history, we select one tracked user every hour whose account is at least 7 days old.
